Remove typgenTypes files + mentions
