--- packages/core/src/typegenTypes.ts
@@ -1,32 +1,0 @@
-import { EventObject, ParameterizedObject, ProvidedActor } from './types.ts';
-
-export interface StateMachineTypes {
-  events: EventObject;
-  actors: ProvidedActor;
-  actions: ParameterizedObject;
-  guards: ParameterizedObject;
-  delays: string;
-  tags: string;
-  emitted: EventObject;
-}
-
-/**
- * @deprecated
- */
-export interface ResolveTypegenMeta<
-  TEvent extends EventObject,
-  TActor extends ProvidedActor,
-  TAction extends ParameterizedObject,
-  TGuard extends ParameterizedObject,
-  TDelay extends string,
-  TTag extends string,
-  TEmitted extends EventObject = EventObject
-> {
-  events: TEvent;
-  actors: TActor;
-  actions: TAction;
-  guards: TGuard;
-  delays: TDelay;
-  tags: TTag;
-  emitted: TEmitted;
-}
